requirements-azure.txt
# Azure ML deployment requirements for EMS Learning Pipeline
azureml-sdk[notebooks]==1.54.0
azureml-mlflow==1.54.0
mlflow>=2.0.1
pandas>=1.5.0
numpy>=1.20.0
matplotlib>=3.5.0
seaborn>=0.11.0
duckdb>=0.8.0
PuLP>=2.7.0
scikit-learn>=1.0.0
scipy>=1.7.0